containers/model_selection/requirements.txt
ollama
openai
pyyaml
requests
flask
flask-socketio
flask-cors
click
dotenv